% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IC_related.R
\name{ic_get_2vect_cnt_table}
\alias{ic_get_2vect_cnt_table}
\title{Get the counts from 2 vectors}
\usage{
ic_get_2vect_cnt_table(V1, V2, pseudo)
}
\arguments{
\item{V1}{A vector.}

\item{V2}{A vector.}

\item{pseudo}{An integer.}
}
\value{
calculate the joint distributions of \code{V1} and \code{V2} with \code{pseudo} as the pseudo counts
}
\description{
Get the counts from 2 vectors
}
\examples{
ic_get_2vect_cnt_table (c("3","2","1"),c("3","2","1"),5)
}
