\name{mutualinfo-package}
\alias{mutualinfo-package}
\alias{mutualinfo}
\docType{package}
\title{
\packageTitle{mutualinfo}
}
\description{
\packageDescription{mutualinfo}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{mutualinfo}
\packageIndices{mutualinfo}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{mutualinfo}

Maintainer: \packageMaintainer{mutualinfo}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
