% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IC_related.R
\name{recoverDf_from_return_list}
\alias{recoverDf_from_return_list}
\title{Convert list to the format of a dataframe}
\usage{
recoverDf_from_return_list(returnList)
}
\arguments{
\item{returnList}{A number.}
}
\value{
Convert \code{returnList} from list to the format of a dataframe
}
\description{
Convert list to the format of a dataframe
}
\examples{
recoverDf_from_return_list (list("a","b","c"))
}
